* Introduced UI module to isolate and manage UI components separately * Initiated a technical design system * Moved the Networking module for improved project organisation
