+++ MoneyBox/AppDelegate.swift
@@ -9,9 +9,14 @@ import UIKit
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
-    var window: UIWindow? 
+
+    // MARK: - Properties
+
+    var window: UIWindow?
+
+    // MARK: - UIApplicationDelegate
+
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        return true
+        true
     }
 }
-
