--- Networking/Sources/Networking/Base/Authentication.swift
@@ -1,12 +1,0 @@
-//
-//  Authentication.swift
-//  MoneyBox
-//
-//  Created by Zeynep Kara on 18.07.2022.
-//
-
-import Foundation
-
-public final class Authentication {
-    public static var token: String?
-}
