--- Networking/Tests/NetworkingTests/NetworkingTests.swift
@@ -1,10 +1,0 @@
-import XCTest
-@testable import Networking
-
-final class NetworkingTests: XCTestCase {
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct
-        // results.
-    }
-}
