+++ Packages/MBUI/Sources/MBUI/Design System/Atoms/Colors.swift
@@ -1,0 +1,18 @@
+//
+//  File.swift
+//
+//
+//  Created by Filippo Minelle on 21/09/2023.
+//
+
+import UIKit
+
+public extension UIColor {
+
+    static var color1: UIColor { UIColor(named: "c1") ?? UIColor(hex: "57BFB4") }
+    static var color2: UIColor { UIColor(named: "c2") ?? UIColor(hex: "FFFFFF") }
+    static var color3: UIColor { UIColor(named: "c3") ?? UIColor(hex: "96DED9") }
+    static var color4: UIColor { UIColor(named: "c4") ?? UIColor(hex: "E4ECEF") }
+    static var color5: UIColor { UIColor(named: "c5") ?? UIColor(hex: "97B7C2") }
+    static var color6: UIColor { UIColor(named: "c6") ?? UIColor(hex: "071C23") }
+}
