+++ Packages/MBUI/Sources/MBUI/Design System/Atoms/File.swift
@@ -1,0 +1,14 @@
+//
+//  File.swift
+//
+//
+//  Created by Filippo Minelle on 21/09/2023.
+//
+
+public enum Padding {
+    public static let xs = 4.0
+    public static let s = 8.0
+    public static let m = 16.0
+    public static let l = 24.0
+    public static let xl = 32.0
+}
