+++ Packages/MBUI/Tests/MBUITests/UIColor+ExtTests.swift
@@ -1,0 +1,20 @@
+// @testable import MBUI
+// import XCTest
+//
+// final class UIColorExtensionTests: XCTestCase {
+//
+//    func testHexInitialization() {
+//        let redColor = UIColor(hex: "#FF0000")
+//        XCTAssertEqual(redColor, UIColor.red)
+//
+//        // Test hexadecimal color with alpha
+//        let blueColorWithAlpha = UIColor(hex: "#0000FF", alpha: 0.5)
+//        XCTAssertEqual(blueColorWithAlpha.cgColor.alpha, 0.5, accuracy: 0.01)
+//    }
+//
+//    func testInvalidHexInitialization() {
+//        // Test with an invalid hexadecimal string
+//        let invalidHexColor = UIColor(hex: "ThisIsNotAValidHex")
+//        XCTAssertEqual(invalidHexColor, UIColor.clear)
+//    }
+// }
