+++ Packages/Networking/Sources/Networking/Base/Authentication.swift
@@ -1,0 +1,12 @@
+//
+//  Authentication.swift
+//  MoneyBox
+//
+//  Created by Zeynep Kara on 18.07.2022.
+//
+
+import Foundation
+
+public enum Authentication {
+    public static var token: String?
+}
