+++ Packages/Networking/Tests/NetworkingTests/NetworkingTests.swift
@@ -1,0 +1,10 @@
+@testable import Networking
+import XCTest
+
+final class NetworkingTests: XCTestCase {
+    func testExample() throws {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct
+        // results.
+    }
+}
